Changes to gene annotation.
I decided to simplify things and just make a gene annotation file. This
file contains everything from FBgn to symbol mapping and chromosome
name and gene length
