+++ Snakefile
@@ -1,6 +1,3 @@
-import yaml
-from pathlib import Path
-
 import pandas as pd
 
 from larval_gonad.config import read_config
@@ -14,8 +11,7 @@ tag = common_config['tag']
 rule targets:
     input:
         expand('references/{assembly}-all-{tag}.modFix.gtf', assembly=assembly, tag=tag),
-        expand('references/fbgn2symbol-{assembly}-all-{tag}.feather', assembly=assembly, tag=tag),
-        expand('references/fbgn2chrom-{assembly}-all-{tag}.feather', assembly=assembly, tag=tag),
+        expand('references/gene_annotation_{assembly}_{tag}.feather', assembly=assembly, tag=tag)
 
 
 rule fix_gtf:
@@ -28,29 +24,15 @@ rule fix_gtf:
     """
 
 
-rule fbgn2symbol:
+rule gene_annotation:
     input: rules.fix_gtf.output[0]
-    output: 'references/fbgn2symbol-{assembly}-all-{tag}.tsv'
-    script: 'scripts/fbgn2symbol.py'
+    output: 'references/gene_annotation_{assembly}_{tag}.tsv'
+    script: 'scripts/gene_annotation.py'
 
 
-rule fbgn2symbol_feather:
-    input: rules.fbgn2symbol.output[0]
-    output: 'references/fbgn2symbol-{assembly}-all-{tag}.feather'
+rule gene_annotation_feather:
+    input: rules.gene_annotation.output[0]
+    output: 'references/gene_annotation_{assembly}_{tag}.feather'
     run:
         df = pd.read_csv(input[0], sep='\t', keep_default_na=False, na_values=['99999999999']).sort_values('FBgn').reset_index(drop=True)
-        df.to_feather(output[0])
-
-
-rule fbgn2chrom:
-    input: rules.fix_gtf.output[0]
-    output: 'references/fbgn2chrom-{assembly}-all-{tag}.tsv'
-    script: 'scripts/fbgn2chrom.py'
-
-
-rule fbgn2chrom_feather:
-    input: rules.fbgn2chrom.output[0]
-    output: 'references/fbgn2chrom-{assembly}-all-{tag}.feather'
-    run:
-        df = pd.read_csv(input[0], sep='\t').sort_values('FBgn').reset_index(drop=True)
         df.to_feather(output[0])
